custom_components/rainpoint/switch.py
"""Switch entities for RainPoint integration."""

import logging

from homeassistant.config_entries import ConfigEntry
from homeassistant.core import HomeAssistant
from homeassistant.helpers.entity_platform import AddEntitiesCallback

from .const import DEBUG_WORKER_URL, DOMAIN
from .coordinator import RainPointCoordinator
from .hub_entities import RainPointHubBroadcastSwitch

_LOGGER = logging.getLogger(__name__)


async def async_setup_entry(
    hass: HomeAssistant,
    entry: ConfigEntry,
    async_add_entities: AddEntitiesCallback,
) -> None:
    """Set up RainPoint switch entities."""
    data = hass.data[DOMAIN][entry.entry_id]
    coordinator: RainPointCoordinator = data["coordinator"]

    entities = []

    # Hub broadcast switches
    hubs_cfg = coordinator.data.get("hubs", [])
    hubs_dict = {str(hub.get("hid", i)): hub for i, hub in enumerate(hubs_cfg)} if isinstance(hubs_cfg, list) else hubs_cfg

    for _hub_key, hub_info in hubs_dict.items():
        entities.append(RainPointHubBroadcastSwitch(coordinator, hub_info))

    # Only register the debug switch when the worker URL is configured
    if DEBUG_WORKER_URL:
        from .debug import RainPointDebugSwitchEntity

        debug_switch = RainPointDebugSwitchEntity(hass, coordinator, entry)
        entities.append(debug_switch)

    _LOGGER.info("Added %d switch entities", len(entities))
    async_add_entities(entities)
