requirements-test.txt
pytest-homeassistant-custom-component>=0.13.310
pytest-asyncio>=0.23
pytest-cov>=4.0.0
